e2e: Rename testcontext.TestContext to test.Context
Repeating the same names in the package name and types is anti-pattern.

I think this a really good name for our use case. The only issue is not
confusing it with context.Context, but since we always use the package
name test.Context is clearly not a context.Context.

The test package can keep other testing helpers that today are in the
generic util package.
