+++ e2e/test/context.go
@@ -1,7 +1,7 @@
 // SPDX-FileCopyrightText: The RamenDR authors
 // SPDX-License-Identifier: Apache-2.0
 
-package testcontext
+package test
 
 import (
 	"testing"
@@ -11,12 +11,12 @@ import (
 	"github.com/ramendr/ramen/e2e/workloads"
 )
 
-type TestContext struct {
+type Context struct {
 	Workload workloads.Workload
 	Deployer deployers.Deployer
 }
 
-func (c *TestContext) Deploy(t *testing.T) {
+func (c *Context) Deploy(t *testing.T) {
 	t.Helper()
 
 	if err := c.Deployer.Deploy(c.Workload); err != nil {
@@ -24,7 +24,7 @@ func (c *TestContext) Deploy(t *testing.T) {
 	}
 }
 
-func (c *TestContext) Undeploy(t *testing.T) {
+func (c *Context) Undeploy(t *testing.T) {
 	t.Helper()
 
 	if err := c.Deployer.Undeploy(c.Workload); err != nil {
@@ -32,7 +32,7 @@ func (c *TestContext) Undeploy(t *testing.T) {
 	}
 }
 
-func (c *TestContext) Enable(t *testing.T) {
+func (c *Context) Enable(t *testing.T) {
 	t.Helper()
 
 	if err := dractions.EnableProtection(c.Workload, c.Deployer); err != nil {
@@ -40,7 +40,7 @@ func (c *TestContext) Enable(t *testing.T) {
 	}
 }
 
-func (c *TestContext) Disable(t *testing.T) {
+func (c *Context) Disable(t *testing.T) {
 	t.Helper()
 
 	if err := dractions.DisableProtection(c.Workload, c.Deployer); err != nil {
@@ -48,7 +48,7 @@ func (c *TestContext) Disable(t *testing.T) {
 	}
 }
 
-func (c *TestContext) Failover(t *testing.T) {
+func (c *Context) Failover(t *testing.T) {
 	t.Helper()
 
 	if err := dractions.Failover(c.Workload, c.Deployer); err != nil {
@@ -56,7 +56,7 @@ func (c *TestContext) Failover(t *testing.T) {
 	}
 }
 
-func (c *TestContext) Relocate(t *testing.T) {
+func (c *Context) Relocate(t *testing.T) {
 	t.Helper()
 
 	if err := dractions.Relocate(c.Workload, c.Deployer); err != nil {
